requirements.txt
anyio==4.7.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
joblib==1.4.2
lightgbm==4.5.0
numpy==2.2.0
python-telegram-bot==21.8
scikit-learn==1.6.0
scipy==1.14.1
sniffio==1.3.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
